refactor: clean up vsock proto command tests
