+++ crates/vsock-proto/src/payloads/write_file.rs
@@ -1,4 +1,4 @@
-use crate::ProtocolError;
+use crate::error::ProtocolError;
 use crate::read::{read_u8_at, read_u16_at, read_u32_at};
 use crate::wire::{WRITE_FILE_FLAG_APPEND, WRITE_FILE_FLAG_SUDO};
